+++ src/ErrorHookReturn.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace PwJsonApi;
+
+/**
+ * Event to return when an error occurs
+ */
+class ErrorHookReturn extends RequestHookReturn
+{
+  /** Exception that was thrown */
+  public ApiException $exception;
+
+  /** Error response (reference to $exception->response) */
+  public Response $response;
+}