+++ src/exceptions/ApiException.php
@@ -10,24 +10,6 @@ class ApiException extends \Exception
   /** Response */
   public readonly Response $response;
 
-  /** Request */
-  public Request $request;
-
-  /** ProcessWire URL hook event */
-  public \ProcessWire\HookEvent $event;
-
-  /** Request endpoint */
-  public Endpoint $endpoint;
-
-  /** Request service */
-  public Service $service;
-
-  /** List of all parent services */
-  public ServiceList $services;
-
-  /** API instance */
-  public Api $api;
-
   /**
    * Constructor
    *